Move function for determining result into VoteResultIcon
