+++ src/components/common/DataUseVoteSummary/VoteResultContainer.js
@@ -1,6 +1,4 @@
-import VoteResultIcon from './VoteResultIcon';
 import VoteResultLabel from './VoteResultLabel';
-import { isEmpty, filter, isNil } from 'lodash/fp';
 import { h, div } from 'react-hyperscript-helpers';
 
 //helper function to generate keys for rendered elements
@@ -8,47 +6,16 @@ const convertLabelToKey = (label) => {
   return label.split(' ').join('-');
 };
 
-const determineUnanimousVoteResult = ({votes = []}) => {
-  const filteredVotes = filter((vote) => !isNil(vote.vote))(votes);
-  if (isEmpty(filteredVotes)) {
-    return 'underReview';
-  }
-  const voteCount = filteredVotes.length;
-
-  let voteTally = {
-    true: 0,
-    false: 0,
-  };
-
-  filteredVotes.forEach((vote = {}) => {
-    voteTally[vote.vote] += 1;
-  });
-
-  if (voteTally.true === voteCount) {
-    return true;
-  } else if (voteTally.false === voteCount) {
-    return false;
-  } else if (voteTally.true + voteTally.false === voteCount) {
-    return 'mixed';
-  } else {
-    return 'underReview';
-  }
-};
-
 export default function VoteResultContainer({
   finalVotes = [],
   label,
   additionalLabelStyle = {},
 }) {
-  const baseContainerStyle = {
-    width: '15%',
-  };
-  const hyphenatedKey = convertLabelToKey(label);
 
-  const result = determineUnanimousVoteResult({votes: finalVotes});
+  const hyphenatedKey = convertLabelToKey(label);
   return div(
     {
-      style: baseContainerStyle,
+      style: { width: '10%'},
       key: hyphenatedKey,
       className: 'vote-summary-container',
     },
@@ -57,7 +24,7 @@ export default function VoteResultContainer({
         propKey: hyphenatedKey,
         label,
         additionalLabelStyle,
-        result
+        finalVotes
       })
     ]
   );
